Task-1/websocket/server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[silero,websocket,google]>=0.0.82
cartesia>=1.1.3
aiohttp>=3.9.0
websockets>=12.0
